feat: add databaseOps module feat: change functions in databaseOps to all be asynchronous
