+++ app.js
@@ -1,24 +1,12 @@
 const { App } = require("@slack/bolt");
 const express = require("express");
+const databaseOps = require("./lib/databaseOps");
 
 const webserver = express();
-var mysql = require("mysql2");
 
-var con = mysql.createConnection({
-  host: process.env.MYSQL_HOSTNAME,
-  user: process.env.MYSQL_USER,
-  password: process.env.MYSQL_PASSWORD,
-  database: process.env.MYSQL_DATABASE,
-});
-
-con.connect(function (err) {
-  if (err) throw err;
-  console.log("Connected!");
-  // con.query('SHOW DATABASES LIKE test', function (err, result) {
-  //   if (err) throw err;
-  //   console.log("Result: " + result);
-  // });
-});
+var con = databaseOps.connectToDB();
+databaseOps.setupDB(con);
+console.log("Connected!");
 
 webserver.get("/healthz", async (req, res) => {
   const status_checks = {};
