+++ lib/sendMessage.js
@@ -1,0 +1,5 @@
+module.exports = {
+  reportPings: (pings) => {
+    return "total pings from you: " + pings;
+  },
+};
